add exunitedcup.com
This is a fake clone website of the real united pro series website at united.gg (found in the linktree listed in the bio of their socials https://twitter.com/UNITEDProSeries / https://www.instagram.com/unitedproseries/). The fake website is a basic steam account scam that attempts to get you to login with steam to vote on a team.